requirements.txt
torch --index-url https://download.pytorch.org/whl/cu124
nltk
pyyaml
python-box
pathlib
dvc
mlflow
dagshub
